Use os.Executable in favor of osext.Executable
Since Go 1.8 os.Executable is part of the stdlib, so osext can be dropped
from dependencies.

That change is not tested, because a test requires to create a s3 bucket
and a fake binary. That said I'm confident because a recent change in
osext invokes os.Executable if the library is compiled under Go 1.8.